WIP to separate helpers into submodule -- need to fix test and method design
