+++ foundry/__init__.py
@@ -1,3 +1,4 @@
 from .foundry import Foundry  # noqa F401 (import unused)
 from . import models  # noqa F401 (import unused)
 from . import https_download  # noqa F401 (import unused)
+from . import https_upload  # noqa F401 (import unused)
