+++ tests/test_foundry.py
@@ -11,6 +11,7 @@ import mdf_toolbox
 import pandas as pd
 from mdf_forge import Forge
 from foundry import Foundry
+from foundry.https_upload import _upload_to_endpoint
 from dlhub_sdk import DLHubClient
 from mdf_connect_client import MDFConnectClient
